Add files via upload
 It works by creating a two-dimensional array to represent the game board, and then adding event listeners to each of the game boxes to detect when a player clicks on one. When a player clicks on a box, the code checks to see if the box is empty. If it is, the code places the player's symbol in the box and updates the game board. The code then checks to see if the player has won the game. If they have, the code disables all of the game boxes and displays a message indicating that the player has won. If the player has not won, the code swaps turns to the other player.

The code also has a function to check for a tie. A tie occurs when all of the game boxes are filled and no player has won. If the code detects a tie, it disables all of the game boxes and displays a message indicating that the game is a tie.

Here is a short description of each function in the code:

initGame(): This function initializes the game by setting the current player to "X", clearing the game board, and enabling all of the game boxes.
swapTurn(): This function swaps turns to the other player.
checkGameOver(): This function checks to see if there is a winner or a tie.
handleClick(): This function is called when a player clicks on a game box. It checks to see if the box is empty and, if it is, it places the player's symbol in the box and updates the game board. It then checks to see if the player has won or if there is a tie.
To play the game, simply open the HTML file in a web browser and click on the game boxes to place your symbols. The first player to get three of their symbols in a row, column, or diagonal wins the game.